Move mock data modules
