+++ src/stream-deck/__mocks__/events/index.ts
@@ -1,0 +1,2 @@
+export * from './did-receive-global-settings';
+export * from './did-receive-settings';
