+++ src/stream-deck/__mocks__/registration/index.ts
@@ -1,0 +1,2 @@
+export * from './action-info';
+export * from './info';
